+++ tests/pipeline/test_generate.py
@@ -11,7 +11,8 @@ from pytest_snapshot.plugin import Snapshot
 from typer.testing import CliRunner
 
 import kpops.api as kpops
-from kpops.cli.main import FilterType, app
+from kpops.api.options import FilterType
+from kpops.cli.main import app
 from kpops.components.base_components.kafka_connector import KafkaSinkConnector
 from kpops.components.base_components.pipeline_component import PipelineComponent
 from kpops.const.file_type import PIPELINE_YAML, KpopsFileType
